Looped and Added Front-End Instructions
I added a while loop that loops through the input and output. To exit the loop, the user can type in either "quit" (any case) or leave the line blank. This is explained at the start, second line on the front end. Above that, a line explains the purpose of the program.